// src/api-chain/api/organization.service.ts
export namespace DeleteOrganization {
    /**
     * Parameter map for deleteOrganization.
     */
    export interface PartialParamMap {
      ChainOrganization: ChainOrganization;
    }

    /**
     * Enumeration of all parameters for deleteOrganization.
     */
    export enum Parameters {
      ChainOrganization = 'ChainOrganization'
    }

    /**
     * A map of tuples with error name and `ValidatorFn` for each parameter of deleteOrganization
     * that does not have an own model.
     */
    export const ParamValidators: {[K in keyof DeleteOrganization.PartialParamMap]?: [string, ValidatorFn][]} = {
    };
}

/**
 * Namespace for getOrganization.
 */
export namespace GetOrganization {
    /**
     * Parameter map for getOrganization.
     */
    export interface PartialParamMap {
      /**
       * database id of the organization
       */
      dbId: string;
    }

    /**
     * Enumeration of all parameters for getOrganization.
     */
    export enum Parameters {
      /**
       * database id of the organization
       */
      dbId = 'dbId'
    }

    /**
     * A map of tuples with error name and `ValidatorFn` for each parameter of getOrganization
     * that does not have an own model.
     */
    export const ParamValidators: {[K in keyof GetOrganization.PartialParamMap]?: [string, ValidatorFn][]} = {
      dbId: [
              ['required', Validators.required],
      ],
    };
}
// ...
export namespace ListOrganizations {
    /**
     * Parameter map for listOrganizations.
     */
    export interface PartialParamMap {
      /**
       * sort order ASC or DESC
       */
      sort?: any;
      /**
       * query limit
       */
      limit?: number;
      /**
       * query offset
       */
      offset?: number;
    }

    /**
     * Enumeration of all parameters for listOrganizations.
     */
    export enum Parameters {
      /**
       * sort order ASC or DESC
       */
      sort = 'sort',
      /**
       * query limit
       */
      limit = 'limit',
      /**
       * query offset
       */
      offset = 'offset'
    }

    /**
     * A map of tuples with error name and `ValidatorFn` for each parameter of listOrganizations
     * that does not have an own model.
     */
    export const ParamValidators: {[K in keyof ListOrganizations.PartialParamMap]?: [string, ValidatorFn][]} = {
      sort: [
      ],
      limit: [
      ],
      offset: [
      ],
    };
}